requirements.txt
PyQt6>=6.4.0
anthropic>=0.3.0
slack-sdk>=3.19.0
python-dotenv>=0.19.0
pytz>=2021.3
structlog>=21.1.0
prometheus-client>=0.11.0
SQLAlchemy>=2.0.27
PyYAML>=6.0.1
typing-extensions>=4.9.0
pydantic>=2.0.0,<3.0.0 